Poisson Disc Sampling
Created a script for poisson disc sampling. This algortihm prevents objects from being within a certain range of each other. This could be used to generate the galaxy to make sure that planets and other spacial objects to not collide. It has yet to be optimised with the burst compiler.