Basic Triplanar Terrain Shader
Created the beginnings of a very basic triplanar shader to the map generation. This allows voxels to be textured within the game and also creates more interesting terrain as the texture of a face is dependent on the facing direction. For example dirt is on the side of voxels whereas grass is on top. Also assigned UV coordinates to the mesh to allow this to happen. The frequency value in the inspector now actually affects the terrain and is not hardcoded. The shader needs to be improved however, as no matter the depth or whether its underground, grass will always be on voxels facing upwards.