Galaxy Additions / Improvements
Added gas giant planets and materials for different types of planets. Reduced size of some objects such as asteroids to reflect their actual scale. The next layer's starting distance in the solar system is now determined by the previous ending position and the number of that item in the solar system. This needs tweaking with the asteroid belt in particular as it can collide with the terrestrial planets. Also edited the poisson disc sampler so that the multi sampler takes in an array of ring inner and outer radii.