Cargo Container
Created a basic blockout for the cargo container. This is a common component that will slot onto expansion slots on ships. Also created expansion slot models (M and F) for where other expansions such as weapon bays can be added to a port on a ship and these will used in other models in particular to make sure dimensions of objects match up.