Substance Plugin Imported
Imported Allegorithmic's substance plugin. This plugin allows substances (high res texture maps that look realistic and make use of physically based lighting) to be added to Unity. The addition of this plugin will mean the project will be converted to HDRP relatively soon in order to make use of these features.