Cargo Container Basic Colouring
The Cargo Container can now be coloured within Unity by changing the basemap albedo colour within the material. This solution is not ideal as the text does not look right. The ideal solution would be to rebuild the textures in Substance Designer and remap the UVs in blender to match. Another possible route is to have static colours for the most part and only have certain elements of each model be colourised.