Cleaned Up Project
Moved external packages to a packages folder for easy access. Removed scripts which were incomplete, messy or generally needed a rewrite, or are not ready to be integrated with the current player systems. The multiplayer also has a bug where movement is not syncing. This will have an issue created and will be fixed with high priority.