Expansion Connector Improvements
Improved the Expansion Connector so that the M port reflects the F port. The M port now has a back plate on it which can be modelled into a ship or any other object. Also created a cargo hauler ship file which demonstrates the cargo container attached to the M port where a ship would be.