Mining Ship 1 Blockout
Blocked out in blender a potential design for an asteroid mining ship. It features a small cargo capacity ideal for quick movement. It has two main thrusters on the back of the ship, attached either side. It has a collector for ores attached to the bottom, ready for direct placement into the cargo containers. Mining Lasers and weapon slots shall be added to the model once the ports / connectors for large components has been modelled.