Networked Ship Spawning
Ship Spawning has been networked. The base system has been vastly reworked to work with Mirror Networking. This means that the adding and removing component system will need to be reworked in order to function correctly. This solution also does not sync ship spawns if a player joins after they have been spawned, so this will require a rework; an issue on Github will be created for this.