Fixed #2
Fixed the issue where the parent transform was not syncing across to newly joined clients. Created a ship controller script which will be used for ship movement. This causes an error to be created when the server ends however as a network behaviour script seems to cause issues when on the same component as a mesh filter / renderer. This means the ship graphical elements shall have to be split up from the logic in order to resolve this issue.