Preparation for Component Syncing
Preparation for component syncing by resolving all errors up to this point. Reorganised the prefab folder. Adding and removing components only works for the host player currently as the arrays for component count are not being set on the clients. Also made it so only one ship can be spawned from a single instance / spawner.