Networked Ship Component Setup
All clients now have the ship component array setup synced, along with the slots associated with them. This does not include adding or removing components just yet. This system works for new clients who have just joined and clients who were in the game when the ship was spawned. In future the ship asset will probably be needed to be networked as well as it is currently running off of the prefab which is spawned on the server.