Cargo Hauler Imported
Imported the newly textured cargo hauler ship into Unity. It now has a prefab and can be used instead of the mining ship within the game. The mining ship has also been reimported in order to fix the rotation of the ship. As previously suspected, the ship asset in use will have to be networked in order to display the correct ship to every client.