Networked Ship Selection
Ship Selection has been implemented. This means that when players join the game they can select the ship they would like; currently this selection is through the method of numbers on the keyboard but UI will be developed for this. This system works by sending a spawn index over the network rather than the larger and less efficient method of sending the entire shipAsset. This is better as the clients will have the ship assets locally, so there is no point in sending over the whole data structure over the network.