Ship Movement - Rolling
Added rolling to the ship controller. This allows the player to roll the ship using the Q and E keys. Also moved the pivot point of the ship models to allow this to happen without rotating strangely around the origin, rather than the ship mesh itself. Removed the unnecessary rigidbody on the ship instance as this system is not used. Organised the ship asset movement variables into categories / headers.