Networked Ship Camera
The ship camera setup function has been networked so that it works on clients. Also made it so the player object is disabled when they enter the ship in order to prevent the player moving while the ship is moving, seperated. The player object is currently disabled when they interact which causes an error with the resync command when E is pressed. This will require a major rework of the player connection system as it is dependent on the player object not being destroyed, something that should be allowwed. An issue will be created on Github for this.