Ship Movement Disable
Ship movement can now be enabled and disabled via a boolean value. This value is false by default, meaning the ship will no longer move when the player is on the ground and can move when the player interacts with the ship. This will be reversed when a certain key is pressed to exit the ship.