Exiting Ships
Ships can now be exited by pressing G. This is because if F is used it still registers the keydown from the previous frame when the player interacts with the ship to enter it. This will require a rework of the interaction system and general input. Every ship prefab now needs an exit point object as the first child of the ship controller object.