Ship Player Object Check
Created several checks for whether a player can enter a ship, These checks are subject to change when #4 has been designed and implemented. Currently, only the owner of the ship can enter and pilot it, through the Mirror Authority System, although a check exists to prevent another entering the ship if another player is already inside it.