Position Map Psuedocode
Improved the programming ship design document so that it contains a workflow guide to how #6 should be implemented. Several files relating the thrusters have been changed and renamed. The cargo ship has had a particle map generated using the current system which raises the idea of a submesh system as mentioned in the design document workflow step 3.