Mesh Extension Utility Class
Created a Utility Script to separate a mesh when given a submesh id. This allows the position map utility script to calculate the index of a certain material in the mesh renderer and then use this as the input of the mesh splitter script. This allows the user to inpu the whole mesh and the material rather than having to create separate models for the point cache. This mesh extension script is used in the creation of #6.