Surface Area Mesh Function
Changed the function used for calculating the amount of pixels that should be dedicated the an area using a function that calculates the surface area of the mesh. This is a better approach as the spawn points are going to be calculated using a random point on each triangle of the mesh, hence the use of surface area rather than volume. This will be used in the creation of #6.