Better Pixel Count Dedication Calculation + Mesh Combination
Changed the way how the even distrubution system works in #6. Instead of working by mesh, it now works by calculating the surface area of the triangles as an array, rather than the surface area of the mesh. All meshes that have the entered material are merged into one mesh and the number of pixels to dedicate to each triangle based on the surface area of each triangle in the mesh is calculated.