Position Mapping for Particle Utility
Implemented a very basic, but semi-functional position mapper. The function to generate the position maps for #6 needs to be reworked heavily as some triangles do not generate at all and there seems to be a heavy bias of particles towards the edges of a triangle. Mapping the positions to the actual texture is now complete however, with the mesh combining proven to be fully functional and not the cause of the issue.