Implemented Particle Position Map Editor #6
The custom editor has finally been completed and fully implemented. This is a major step forward for the particle system. Their will be numerous improvements to this system as it's used more. One issue that will be created on github is how the last triangle in the mesh does not seem to have any pixel data generated for it. A temporary workaround has been implemented to prevent the editor from crashing but this will be rectified shortly.