Ship Particle Spawn Rate and Velocity Based on Ship Velocity
The spawn rate for particles and the velocity of them is now determined based on the velocity of the ship. This system is also networked. A small visual error occurs when the velocity is too great and this causes the particles to look "stagnated" or grouped into how many spawn in that frame. This can be fixed by adding a tolerance value to the particle position map generator, which will add or remove a random X Y Z value from each position to create some randomness which should help reduce this grouping visual artifact. The values for movement still need to be tweaked to try and make movement feel better.