Planet Performance Improvements (Culling)
Implemented functions for culling parts of the planet that cannot be seen by the player. This is known as occlusion culling. Also updated the LOD values to be an array rather than a dictionary. The values are also much more realistic for the planet size. Better numbers will be calculated and then multiplied up by the size of the planet to allow different size planets. Currently planets can not be smaller than 1000 due to strange glitches. This may be fixed by the LOD value multiplier however.